sam's class ranges heal: Rune joins the vendored tree's 21-bit era
The reported bug: sam -d's ,x/[a-z]+es/ c/X/ silently no-ops while
,x/courses/ works (the suite's literal-only sam test never caught it).
Narrowed by measurement: enumerated classes, negation, closures and
alternation all matched; every x-y RANGE never did.

Root cause, three layers deep, all in the shim (vendored sources
untouched): u.h declared Rune as unsigned short ("the 4th edition's"),
but the vendored snapshot is the LATE 4th edition — libc.h's Runemax
is 0x10FFFF, UTFmax 4, rune.c does surrogates. The one construct that
needs Runemax to round-trip through a Rune is sam's class-range
sentinel (regexp.c: classp[n] = Runemax, matched by *p == Runemax) —
16-bit storage truncates it to 0xFFFF and the compare never fires, so
[a-z] degrades to the literal set {0xFFFF,'a','z'}.

The fix cascaded twice, each stop measured: Rune -> unsigned int broke
L"..." literals (wasm32's wchar_t is a SIGNED int and clang refuses
initialising an unsigned array from it; -fshort-wchar retired, P9CC now
carries -Xclang -fwchar-type=int -Xclang -fno-signed-wchar); and then
wc counted zero words — lib9.c's _runebsearch, OUR platform helper
behind vendored runetype.c, had unsigned short hardcoded and
half-strided every 32-bit classification table. That mismatch links
silently: wasm promotes u16 and u32 alike to i32, so the wrong C
signature produces the right wasm signature. RESEARCH §9.4 records all
three findings.

The suite gains the pinning test — ,x/[a-z]+es/ g/co/ c/COURSES/ must
land — and the floor's cushion grows: 133 PASS on all three hosts
(oracle, wasmtime native, demo lineage); the frozen oracle stayed
untouched and green throughout. Status lines advance to 133; dated
history keeps its 132s.

+++ userspace/libc/lib9.c
@@ -268,9 +268,13 @@ int setjmp(long *env){
 	_sjbuf((int)_asydata);
 	return _setj((int)env);
 }
-/* the rune-range binary search runetype.c leans on */
-unsigned short *_runebsearch(unsigned short c, unsigned short *t, int n, int ne){
-	unsigned short *p;
+/* the rune-range binary search runetype.c leans on — element type MUST
+ * match the platform Rune (u.h: unsigned int, the 21-bit era): the callers
+ * pass Rune tables, and a narrower type here half-strides them silently
+ * (wasm promotes both to i32, so the link never complains — measured
+ * 2026-08-29 as wc counting zero words) */
+unsigned int *_runebsearch(unsigned int c, unsigned int *t, int n, int ne){
+	unsigned int *p;
 	int m;
 
 	while(n > 1){
